+++ lib/syntax_suggest/api.rb
@@ -227,6 +227,7 @@ require_relative "lex_all"
 require_relative "code_line"
 require_relative "code_block"
 require_relative "block_expand"
+require_relative "mini_stringio"
 require_relative "priority_queue"
 require_relative "unvisited_lines"
 require_relative "around_block_scan"
